Initial implementation of DynamicEvent
