+++ Cargo.toml
@@ -10,8 +10,8 @@ log = { version = "0.4.6", features = ["release_max_level_info"] }
 futures = "0.1.26"
 tokio = "0.1.19"
 warp = { git = "https://github.com/seanmonstar/warp.git"}
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.39"
+serde = { version = "1.0.105", features = ["derive"] }
+serde_json = "1.0.50"
 serde_derive = "1.0.90"
 pretty_env_logger = "0.3.0"
 postgres = "0.17.0"
@@ -36,4 +36,5 @@ harness = false
 
 [features]
 default = [ "production" ]
+bench = []
 production = []
