+++ src/messages/event/checked_event/status/application.rs
@@ -1,0 +1,11 @@
+use serde::{Deserialize, Serialize};
+
+#[serde(deny_unknown_fields)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+pub(super) struct Application {
+    name: String,
+    website: Option<String>,
+    vapid_key: Option<String>,
+    client_id: Option<String>,
+    client_secret: Option<String>,
+}
