+++ src/messages/test.rs
@@ -1,3 +1,47 @@
+use super::*;
+use std::collections::HashMap;
+
+#[serde(rename_all = "snake_case")]
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub enum Event {
+    Update,
+    Notification,
+    Delete,
+    FiltersChanged,
+    Announcement,
+    #[serde(rename(
+        serialize = "announcement.reaction",
+        deserialize = "announcement.reaction"
+    ))]
+    AnnouncementReaction,
+    #[serde(rename(serialize = "announcement.delete", deserialize = "announcement.delete"))]
+    AnnouncementDelete,
+    Conversation,
+}
+
+#[serde(rename_all = "snake_case", tag = "event")]
+#[derive(Deserialize, Debug, Clone, PartialEq)]
+pub struct Msg {
+    event: Event,
+    queued_at: Option<i64>,
+}
+#[test]
+fn parse_redis_msg_to() {
+    let input = REDIS_MSG_EVENT_TXT;
+    let msg: Msg = serde_json::from_str(input).expect("TODO");
+    dbg!(&msg);
+    let raw_txt = match (msg.event, msg.queued_at) {
+        (Event::Update, Some(n)) => {
+            &input[r#"{"event":"update","payload":"#.len()
+                ..input.len() - format!(",\"queued_at\":{}", n).len()]
+        }
+        _ => unimplemented!(),
+    };
+    dbg!(raw_txt);
+    panic!("TODO");
+}
+
+fn parse_redis_msg_to_old() {}
 // TODO: Revise these tests to cover *only* the RedisMessage -> (Timeline, Event) parsing
 //     use super::*;
 //     use crate::{
@@ -507,3 +551,5 @@
 //         assert_eq!(rest, String::new());
 //         Ok(())
 //     }
+
+const REDIS_MSG_EVENT_TXT: &str = r#"{"event":"update","payload":{"id":"102775370117886890","created_at":"2019-09-11T18:42:19.000Z","in_reply_to_id":null,"in_reply_to_account_id":null,"sensitive":false,"spoiler_text":"","visibility":"unlisted","language":"en","uri":"https://mastodon.host/users/federationbot/statuses/102775346916917099","url":"https://mastodon.host/@federationbot/102775346916917099","replies_count":0,"reblogs_count":0,"favourites_count":0,"favourited":false,"reblogged":false,"muted":false,"content":"<p>Trending tags:<br><a href=\"https://mastodon.host/tags/neverforget\" class=\"mention hashtag\" rel=\"nofollow noopener\" target=\"_blank\">#<span>neverforget</span></a><br><a href=\"https://mastodon.host/tags/4styles\" class=\"mention hashtag\" rel=\"nofollow noopener\" target=\"_blank\">#<span>4styles</span></a><br><a href=\"https://mastodon.host/tags/newpipe\" class=\"mention hashtag\" rel=\"nofollow noopener\" target=\"_blank\">#<span>newpipe</span></a><br><a href=\"https://mastodon.host/tags/uber\" class=\"mention hashtag\" rel=\"nofollow noopener\" target=\"_blank\">#<span>uber</span></a><br><a href=\"https://mastodon.host/tags/mercredifiction\" class=\"mention hashtag\" rel=\"nofollow noopener\" target=\"_blank\">#<span>mercredifiction</span></a></p>","reblog":null,"account":{"id":"78","username":"federationbot","acct":"federationbot@mastodon.host","display_name":"Federation Bot","locked":false,"bot":false,"created_at":"2019-09-10T15:04:25.559Z","note":"<p>Hello, I am mastodon.host official semi bot.</p><p>Follow me if you want to have some updates on the view of the fediverse from here ( I only post unlisted ). </p><p>I also randomly boost one of my followers toot every hour !</p><p>If you don't feel confortable with me following you, tell me: unfollow  and I'll do it :)</p><p>If you want me to follow you, just tell me follow ! </p><p>If you want automatic follow for new users on your instance and you are an instance admin, contact me !</p><p>Other commands are private :)</p>","url":"https://mastodon.host/@federationbot","avatar":"https://instance.codesections.com/system/accounts/avatars/000/000/078/original/d9e2be5398629cf8.jpeg?1568127863","avatar_static":"https://instance.codesections.com/system/accounts/avatars/000/000/078/original/d9e2be5398629cf8.jpeg?1568127863","header":"https://instance.codesections.com/headers/original/missing.png","header_static":"https://instance.codesections.com/headers/original/missing.png","followers_count":16636,"following_count":179532,"statuses_count":50554,"emojis":[],"fields":[{"name":"More stats","value":"<a href=\"https://mastodon.host/stats.html\" rel=\"nofollow noopener\" target=\"_blank\"><span class=\"invisible\">https://</span><span class=\"\">mastodon.host/stats.html</span><span class=\"invisible\"></span></a>","verified_at":null},{"name":"More infos","value":"<a href=\"https://mastodon.host/about/more\" rel=\"nofollow noopener\" target=\"_blank\"><span class=\"invisible\">https://</span><span class=\"\">mastodon.host/about/more</span><span class=\"invisible\"></span></a>","verified_at":null},{"name":"Owner/Friend","value":"<span class=\"h-card\"><a href=\"https://mastodon.host/@gled\" class=\"u-url mention\" rel=\"nofollow noopener\" target=\"_blank\">@<span>gled</span></a></span>","verified_at":null}]},"media_attachments":[],"mentions":[],"tags":[{"name":"4styles","url":"https://instance.codesections.com/tags/4styles"},{"name":"neverforget","url":"https://instance.codesections.com/tags/neverforget"},{"name":"mercredifiction","url":"https://instance.codesections.com/tags/mercredifiction"},{"name":"uber","url":"https://instance.codesections.com/tags/uber"},{"name":"newpipe","url":"https://instance.codesections.com/tags/newpipe"}],"emojis":[],"card":null,"poll":null},"queued_at":1568227693541}"#;
