+++ src/redis_to_client_stream/client_agent.rs
@@ -17,7 +17,7 @@
 //! each new client connection (each in its own thread).use super::{message::Message, receiver::Receiver}
 use super::receiver::{Receiver, ReceiverErr};
 use crate::{
-    messages::Event,
+    messages::{CheckedEvent, DynamicEvent, Event},
     parse_client_request::{Stream::Public, Subscription, Timeline},
 };
 use futures::{
@@ -96,29 +96,48 @@ impl futures::stream::Stream for ClientAgent {
         let blocked_domains = &self.subscription.blocks.blocked_domains;
         let (send, block) = (|msg| Ok(Ready(Some(msg))), Ok(NotReady));
 
-        use Event::*;
+        use CheckedEvent::*;
         match result {
-            Ok(Async::Ready(Some(event))) => match event {
-                Update {
-                    payload: status, ..
-                } => match self.subscription.timeline {
-                    _ if status.involves_blocked_user(blocked_users) => block,
-                    _ if status.from_blocked_domain(blocked_domains) => block,
-                    _ if status.from_blocking_user(blocking_users) => block,
-                    Timeline(Public, _, _) if status.language_not_allowed(allowed_langs) => block,
-                    _ => send(Update {
-                        payload: status,
-                        queued_at: None,
-                    }),
-                },
-                Notification { .. }
-                | Conversation { .. }
-                | Delete { .. }
-                | FiltersChanged
-                | Announcement { .. }
-                | AnnouncementReaction { .. }
-                | AnnouncementDelete { .. } => send(event),
-            },
+            Ok(Async::Ready(Some(event))) => {
+                let event = match serde_json::from_str(&event) {
+                    Ok(checked_event) => Event::TypeSafe(checked_event),
+                    Err(e) => {
+                        log::error!(
+                            "Error safely parsing Redis input.\
+  Mastodon and Flodgatt do not \
+                             strictly conform to the same \
+version of Mastodon's API.\n{}",
+                            e
+                        );
+                        let dyn_event: DynamicEvent = serde_json::from_str(&event).expect("TODO");
+                        Event::Dynamic(dyn_event)
+                    }
+                };
+
+                match event {
+                    Event::TypeSafe(Update { payload: s, .. }) => {
+                        match self.subscription.timeline {
+                            Timeline(Public, _, _) if s.language_not(allowed_langs) => block,
+                            _ if s.involves_any(blocked_users, blocked_domains, blocking_users) => {
+                                block
+                            }
+                            _ => send(Event::TypeSafe(Update {
+                                payload: s,
+                                queued_at: None,
+                            })),
+                        }
+                    }
+                    Event::TypeSafe(_) => send(event),
+                    Event::Dynamic(e) if e.event == "update" => match self.subscription.timeline {
+                        Timeline(Public, _, _) if e.language_not(allowed_langs) => block,
+                        _ if e.involves_any(blocked_users, blocked_domains, blocking_users) => {
+                            block
+                        }
+                        _ => send(Event::Dynamic(e)),
+                    },
+                    Event::Dynamic(_) => send(event),
+                }
+            }
             Ok(Ready(None)) => Ok(Ready(None)),
             Ok(NotReady) => Ok(NotReady),
             Err(e) => Err(e),
